requirements.txt
# Partner Delivery - Requirements
# Run: pip install -r requirements.txt

ultralytics>=8.0.0
boto3>=1.26.0
pyyaml>=6.0
opencv-python>=4.8.0
